Frame-align the play cursor, fix pre-transformed depth, add stall probes
RanAudio_VoicePosition returned a fractional-frame byte offset. Real
DirectSound cursors sit on a sample-frame boundary, and the client
subtracts two of them to size a decode request: a 2-byte tail made
ov_read floor length/bytespersample to zero samples and answer
OV_EINVAL (-131) for as long as it was asked. Truncate to whole frames,
and hold to the same guarantee at GetCurrentPosition.

The vertex shader passed a D3DFVF_XYZRHW z straight into gl_Position.
D3D clip z is [0,w] and GL's is [-w,w], so a z of 0 - the near plane -
landed at window depth 0.5, in the middle of whatever the last scene
left in the depth buffer.

Instrumentation, all diagnostic-gated or free:
  - EnterCriticalSection, the event wait and pthread_join report their
    own caller through dladdr when they block for more than 3 seconds
  - threads are named after the routine they run, so comm identifies a
    spinning worker with no debugger on the device
  - RanGL_ProbePixel, RanGL_ProbeState and RanD3D_DumpTexture, for the
    loading screen, which renders outside the frame path the other
    diagnostics hang off

+++ native/shim/platform/audio_mix.cpp
@@ -458,7 +458,12 @@ extern "C" unsigned RanAudio_VoicePosition ( int voice )
             while (pos < 0.0) pos += ringFrames;
         } else if (pos < 0.0) pos = 0.0;
 
-        at = (unsigned) ( pos * frameBytes );
+        //  Whole frames only. A cursor part way through a frame is not a thing
+        //  real hardware reports, and the client subtracts two of these to size
+        //  a decode request: a request that is not a multiple of the frame size
+        //  leaves a tail ov_read can never satisfy, and it answers OV_EINVAL
+        //  for as long as it is asked.
+        at = (unsigned) ( (long long) pos ) * frameBytes;
     }
     pthread_mutex_unlock ( &g_lock );
     return at;
